+++ src/main/java/racing/Main.java
@@ -2,8 +2,8 @@ package racing;
 
 import racing.carRacing.model.Cars;
 import racing.carRacing.model.Count;
-import racing.carRacing.service.RandomMovementStrategy;
-import racing.carRacing.service.Stages;
+import racing.carRacing.model.RandomMovementStrategy;
+import racing.carRacing.model.Stages;
 import racing.carRacing.view.InputView;
 import racing.carRacing.view.OutputView;
